clearing casting warnings
git-svn-id: svn://orca.physics.unc.edu/Orca64@11948 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Data Flow/Fan Out/ORFanOutModel.h
@@ -30,7 +30,7 @@
     int			 lineType;
     NSColor*	 lineColor;
 	NSMutableArray* cachedProcessors;
-	uint32_t cachedProcessorsCount;
+	NSUInteger cachedProcessorsCount;
 }
 
 #pragma mark ¥¥¥Initialization
